+++ src/main/java/com/daklan/controlbudget/rest/repository/EmailRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * <b>The repository class for CRUD operations on Email.</b>
  */
-public interface EmailRepository extends JpaRepository<EmailEntity, Integer> {
+public interface EmailRepository extends JpaRepository<EmailEntity, Long> {
     EmailEntity findByIdentifiedBy(String identifiedBy);
 }
