+++ src/main/java/com/daklan/controlbudget/rest/repository/PersonRepository.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 /**
  * <b>The repository class for CRUD operations on PERSON.</b>
  */
-public interface PersonRepository extends JpaRepository<PersonEntity, Integer> {
+public interface PersonRepository extends JpaRepository<PersonEntity, Long> {
     PersonEntity findByFirstNameAndLastNameAndBirthDate(String firstName, String lastName, LocalDate birthDate);
 }
